reminder/requirements.txt
telethon
asyncpg
sqlalchemy
nest-asyncio
aioschedule
python-dotenv
